# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kitchen Pantry CRM API
  description: |
    A comprehensive RESTful API for managing food service industry relationships, interactions, and sales opportunities. 
    Built on Node.js/Express with Supabase backend integration, supporting iPad-first CRM operations.
  version: 1.0.0
  contact:
    name: Kitchen Pantry CRM Support
    email: support@kitchenpantry.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kitchenpantry.com/v1
    description: Production server
  - url: https://staging-api.kitchenpantry.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /interactions:
    get:
      summary: List interactions
      description: Retrieves interaction history with filtering, sorting, and pagination capabilities.
      tags:
        - Interactions
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of records per page
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: organization_id
          in: query
          description: Filter interactions by organization
          schema:
            type: string
            format: uuid
        - name: contact_id
          in: query
          description: Filter interactions by contact
          schema:
            type: string
            format: uuid
        - name: user_id
          in: query
          description: Filter interactions by user
          schema:
            type: string
            format: uuid
        - name: interaction_type
          in: query
          description: Filter by interaction type
          schema:
            type: string
            enum: [CALL, EMAIL, MEETING, VISIT, TEXT, SOCIAL]
        - name: date_from
          in: query
          description: Filter interactions from specified date
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: Filter interactions to specified date
          schema:
            type: string
            format: date
        - name: outcome
          in: query
          description: Filter by interaction outcome
          schema:
            type: string
            enum: [Positive, Neutral, Negative, Very Positive, Very Negative]
        - name: follow_up_required
          in: query
          description: Filter interactions requiring follow-up
          schema:
            type: boolean
        - name: priority_level
          in: query
          description: Filter by priority level
          schema:
            type: string
            enum: [A, B, C, D]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create interaction
      description: Creates a new interaction record with automatic relationship updates and follow-up scheduling.
      tags:
        - Interactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreateRequest'
      responses:
        '201':
          description: Interaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /interactions/{id}:
    get:
      summary: Get interaction by ID
      description: Retrieves detailed information for a specific interaction.
      tags:
        - Interactions
      parameters:
        - name: id
          in: path
          required: true
          description: Interaction ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update interaction
      description: Updates an existing interaction with complete resource replacement.
      tags:
        - Interactions
      parameters:
        - name: id
          in: path
          required: true
          description: Interaction ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionUpdateRequest'
      responses:
        '200':
          description: Interaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Partially update interaction
      description: Performs partial updates to an interaction, updating only specified fields.
      tags:
        - Interactions
      parameters:
        - name: id
          in: path
          required: true
          description: Interaction ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionPatchRequest'
      responses:
        '200':
          description: Interaction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Delete interaction
      description: Soft deletes an interaction by setting is_deleted = true.
      tags:
        - Interactions
      parameters:
        - name: id
          in: path
          required: true
          description: Interaction ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Interaction deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued by Supabase Auth
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        email:
          type: string
          format: email
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 3
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
    ResponseMetadata:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Organization schemas
...
    Interaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        contact:
          type: object
          properties:
            id:
              type: string
              format: uuid
            full_name:
              type: string
            job_title:
              type: string
        user:
          $ref: '#/components/schemas/User'
        interaction_type:
          type: string
          enum: [CALL, EMAIL, MEETING, VISIT, TEXT, SOCIAL]
        interaction_date:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          minimum: 0
        subject:
          type: string
        description:
          type: string
        outcome:
          type: string
          enum: [Positive, Neutral, Negative, Very Positive, Very Negative]
        outcome_details:
          type: string
        follow_up_required:
          type: boolean
        follow_up_date:
          type: string
          format: date-time
        follow_up_notes:
          type: string
        location:
          type: string
        meeting_attendees:
          type: array
          items:
            type: string
        sentiment_score:
          type: integer
          minimum: 1
          maximum: 5
        priority_level:
          type: string
          enum: [A, B, C, D]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    InteractionCreateRequest:
      type: object
      properties:
        organization_id:
          type: string
          format: uuid
        contact_id:
          type: string
          format: uuid
        interaction_type:
          type: string
          enum: [CALL, EMAIL, MEETING, VISIT, TEXT, SOCIAL]
        interaction_date:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          minimum: 0
        subject:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        outcome:
          type: string
          enum: [Positive, Neutral, Negative, Very Positive, Very Negative]
        outcome_details:
          type: string
        follow_up_required:
          type: boolean
          default: false
        follow_up_date:
          type: string
          format: date-time
        follow_up_notes:
          type: string
        location:
          type: string
        meeting_attendees:
          type: array
          items:
            type: string
        sentiment_score:
          type: integer
          minimum: 1
          maximum: 5
        priority_level:
          type: string
          enum: [A, B, C, D]
        tags:
          type: array
          items:
            type: string
      required:
        - organization_id
        - contact_id
        - interaction_type
        - interaction_date
        - subject
        - outcome
    InteractionUpdateRequest:
      $ref: '#/components/schemas/InteractionCreateRequest'
    InteractionPatchRequest:
      type: object
      properties:
        interaction_type:
          type: string
          enum: [CALL, EMAIL, MEETING, VISIT, TEXT, SOCIAL]
        interaction_date:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          minimum: 0
        subject:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        outcome:
          type: string
          enum: [Positive, Neutral, Negative, Very Positive, Very Negative]
        outcome_details:
          type: string
        follow_up_required:
          type: boolean
        follow_up_date:
          type: string
          format: date-time
        follow_up_notes:
          type: string
        location:
          type: string
        meeting_attendees:
          type: array
          items:
            type: string
        sentiment_score:
          type: integer
          minimum: 1
          maximum: 5
        priority_level:
          type: string
          enum: [A, B, C, D]
        tags:
          type: array
          items:
            type: string
    InteractionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Interaction'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'
    InteractionDetailResponse:
      $ref: '#/components/schemas/InteractionResponse'
    InteractionListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Interaction'
        meta:
          $ref: '#/components/schemas/ResponseMetadata'

    # Opportunity schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              field:
                type: string
              details:
                type: string
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
            retry_after:
              type: integer
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: UNAUTHORIZED
                message: Authentication required
                details: Valid JWT token required in Authorization header
    ForbiddenError:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: FORBIDDEN
                message: Access denied
                details: Insufficient permissions for this resource
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: NOT_FOUND
                message: Resource not found
                details: The requested resource does not exist
    BadRequestError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: BAD_REQUEST
                message: Invalid request format
                details: Request body must be valid JSON
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: VALIDATION_ERROR
                message: Invalid input data
                field: email_primary
                details: Email format is invalid
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            errors:
              - code: INTERNAL_ERROR
                message: Internal server error
                details: An unexpected error occurred
...